Program 001
This is the part of sigma web development course by code with harry and it is among the final javascript exercises.
i have added a jquery library for toasters and used some ai to addd a bit of ai. though the card and other logic was made by me, the input files  logic was simply copied from MDN